+++ CHANGES.md
@@ -39,6 +39,7 @@ Connectors:
 
 -   Made the CSV connector put the description column in the distribution description, not just the dataset one
 -   Fixed: CSW connector does not capture all distributions for some datasources (e.g. TERN)
+-   Fixed: CSW connector does not capture persistent URL or licences for some sources (eg. geoscience australia)
 
 Interoperability: